LF-5216 Move getDeviceType to utils/ and rename the conflicting function from sensors
