+++ packages/webapp/src/util/getDeviceType.ts
